Moved AppsClient to a plugin
Example in README

App client settings section in config
