+++ plugins/clients/apps-client/setup.py
@@ -1,0 +1,56 @@
+import setuptools
+
+version = {}
+with open("../../../engine/src/hopeit/server/version.py") as fp:
+    exec(fp.read(), version)
+
+setuptools.setup(
+    name="hopeit.apps-client",
+    version=version['ENGINE_VERSION'],
+    description="Hopeit Apps Client Plugin",
+    license="Apache 2",
+    long_description=open('README.md').read(),
+    long_description_content_type="text/markdown",
+    author="Leo Smerling and Pablo Canto",
+    author_email="contact@hopeit.com.ar",
+    url="https://github.com/hopeit-git/hopeit.engine",
+    classifiers=[
+        "License :: OSI Approved :: Apache Software License",
+        "Intended Audience :: Developers",
+        "Programming Language :: Python",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Development Status :: 4 - Beta",
+        "Operating System :: POSIX :: Linux",
+        "Operating System :: MacOS :: MacOS X",
+        "Operating System :: Microsoft :: Windows",
+        "Topic :: Internet :: WWW/HTTP",
+        "Topic :: Software Development :: Libraries :: Python Modules",
+        "Framework :: AsyncIO",
+    ],    
+    project_urls={
+        "CI: GitHub Actions": "https://github.com/hopeit-git/hopeit.engine/actions?query=workflow",  # noqa
+        "Docs: RTD": "https://hopeitengine.readthedocs.io/en/latest/",
+        "GitHub: issues": "https://github.com/hopeit-git/hopeit.engine/issues",
+        "GitHub: repo": "https://github.com/hopeit-git/hopeit.engine",
+    },
+    package_dir={
+        "": "src"
+    },
+    packages=[
+        "hopeit.apps_client"
+    ],
+    include_package_data=True,
+    package_data={
+        "hopeit.apps_client": ["py.typed"]
+    },
+    python_requires=">=3.7",
+    install_requires=[
+        f"hopeit.engine[web]=={version['ENGINE_VERSION']}"
+    ],
+    extras_require={
+    },
+    entry_points={
+    }
+)
